# perf: improve performance of Link when navigating to unique routes